golden-paths/h3-innovation/sre-agent-integration/skeleton/requirements.txt
azure-identity>=1.15.0
azure-monitor-query>=1.2.0
openai>=1.0.0
fastapi>=0.110.0
uvicorn>=0.27.0
pydantic>=2.0.0
